JS-2209 Move testing recognition helpers
