+++ packages/analysis/src/jsts/rules/S8998/rule.ts
@@ -24,7 +24,7 @@ import {
   getParameterizedTestFrameworks,
   hasSupportedParameterizedCallback,
   isSupportedParameterizedDeclaration,
-} from '../helpers/mocha-style-test-frameworks.js';
+} from '../helpers/testing/mocha-style-test-frameworks.js';
 import * as meta from './generated-meta.js';
 
 const messages = {
