+++ packages/analysis/src/jsts/rules/helpers/testing/mocha-style-test-frameworks.ts
@@ -23,8 +23,8 @@ import {
   isIdentifier,
   isStaticTemplateLiteral,
   isStringLiteral,
-} from './ast.js';
-import { getFullyQualifiedName, importsOrDependsOnModule } from './module.js';
+} from '../ast.js';
+import { getFullyQualifiedName, importsOrDependsOnModule } from '../module.js';
 
 export const SUPPORTED_TEST_FRAMEWORKS = [
   'bun:test',
